refactor: remove BugFender logging and fix IntelliJ code inspections
  - Remove BugFender dependency and all remote logging infrastructure
  - Replace with standard Android Log for local debugging via Logcat
  - Fix 39 IntelliJ code inspection issues:
    - Remove unused resources (7 colors, 1 drawable)
    - Fix function naming conventions (2 functions)
    - Remove redundant suspend modifiers (6 functions)
    - Fix operator precedence with clarifying parentheses (2 issues)
    - Resolve KDoc unresolved references (2 issues)
    - Simplify redundant code patterns and qualifiers
    - Remove unused imports across affected files
  - Add Android Log mocking for tests with robolectric configuration

+++ app/src/main/java/com/github/radupana/featherweight/service/OneRMService.kt
@@ -73,7 +73,7 @@ class OneRMService {
                     calc
                 }
                 RMScalingType.ISOLATION -> {
-                    val calc = weight * totalRepCapacity.toFloat().pow(0.10f)
+                    val calc = weight * totalRepCapacity.pow(0.10f)
                     Log.d("OneRMService", "Isolation formula: $weight × $totalRepCapacity^0.10 = $calc")
                     calc
                 }
